+++ lib/llm/service/completion_batch.py
@@ -6,8 +6,8 @@ from openai import OpenAI
 from openai.types import Batch
 
 from lib.llm.service.completion import LlmService
-from lib.llm.valueobject.chat import Message, RoleType
-from lib.llm.valueobject.chat_batch import MessageChunk
+from lib.llm.valueobject.completion import Message, RoleType
+from lib.llm.valueobject.completion_batch import MessageChunk
 from lib.llm.valueobject.config import OpenAIGptConfig
