+++ lib/llm/valueobject/completion.py
