+++ llm_chat/models.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.models import User
 from django.db import models
 
-from lib.llm.valueobject.chat import RoleType
+from lib.llm.valueobject.completion import RoleType
 
 
 class ChatLogs(models.Model):
